thoughts on realism and art
thinking more about this assignment, this interaction is about proximity and timing. the cursor’s influence over the central sphere is deliberately complicated by an unintuitive orientation as well as a competing orbit. the speed of the second sphere fluctuates just beyond the central sphere’s path, making it tricky to connect the two. the shifting colors of the canvas, both with time and the cursor, can further obscure these two main spheres. 

a connection correlates with a third, much darker sphere. this one grows and shifts in shading depending on the length and degree of overlap. 

my hope was to pose a game-like question to the user: what is the maximum number of spheres you can see at once? there are actually two more, which come about with the right timing. the “flare” pink sphere fades in and out of floating along the surface of the smaller sphere. within this one, for about .5% of the time, an even tinier sphere orbits. so in total, there could be 5 spheres at one time. 

intrinsic in this mechanic as well as my “messy” aesthetic is this thought of collaborating with the machine. i tried for some time to make one sphere pass in front of another, but what i came out with was this. i thought, “why am i trying to emulate reality?” what the machine is doing is very cool. while i would like to pursue a deep understanding of graphics, i would like to do so with an eye towards embracing the abilities and constraints of the technology. i believe that it is only in putting those constraints forward that the work transcends the race for realism, and becomes art.